+++ src/controllers/api/magicItemController.ts
@@ -1,8 +1,8 @@
 import MagicItem from '../../models/magicItem';
-const { redisClient, escapeRegExp, ResourceList } = require('../../util');
+import { redisClient, escapeRegExp, ResourceList } from '../../util';
 import { Request, Response } from 'express';
 
-export const index = async (req: Request, res: Response, next: any) => {
+export const index = async (req: any, res: any, next: any) => {
   const searchQueries: { name?: any } = {};
   if (req.query.name !== undefined) {
     searchQueries.name = { $regex: new RegExp(escapeRegExp(req.query.name), 'i') };
@@ -32,7 +32,7 @@ export const index = async (req: Request, res: Response, next: any) => {
   }
 };
 
-export const show = async (req: Request, res: Response, next: any) => {
+export const show = async (req: any, res: any, next: any) => {
   try {
     const data = await MagicItem.findOne({ index: req.params.index });
     if (!data) return next();
