+++ src/controllers/api/ruleController.ts
@@ -1,8 +1,8 @@
 import Rule from '../../models/rule';
-const { redisClient, escapeRegExp, ResourceList } = require('../../util');
+import { redisClient, escapeRegExp, ResourceList } from '../../util';
 import { Request, Response } from 'express';
 
-export const index = async (req: Request, res: Response, next: any) => {
+export const index = async (req: any, res: any, next: any) => {
   const searchQueries: { name?: any, desc?: any } = {};
   if (req.query.name !== undefined) {
     searchQueries.name = { $regex: new RegExp(escapeRegExp(req.query.name), 'i') };
@@ -35,7 +35,7 @@ export const index = async (req: Request, res: Response, next: any) => {
   }
 };
 
-export const show = async (req: Request, res: Response, next: any) => {
+export const show = async (req: any, res: any, next: any) => {
   try {
     const data = await Rule.findOne({ index: req.params.index });
     if (!data) return next();
