+++ src/tests/controllers/api/equipmentCategoryController.test.ts
@@ -1,40 +1,38 @@
 const mockingoose = require('mockingoose');
-const { mockRequest, mockResponse, mockNext } = require('../support/requestHelpers');
-const AbilityScore = require('../../models/abilityScore');
-const SimpleController = require('../../controllers/simpleController');
+import { mockRequest, mockResponse, mockNext } from '../../support/requestHelpers';
+import EquipmentCategory from '../../../models/equipmentCategory';
+import EquipmentCategoryController from '../../../controllers/api/equipmentCategoryController';
 
-let response;
-let simpleController;
+let response: any;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
-  simpleController = new SimpleController(AbilityScore);
 });
 
 describe('index', () => {
   const findDoc = [
     {
-      index: 'str',
-      name: 'STR',
-      url: '/api/ability-scores/str',
+      index: 'adventuring-gear',
+      name: 'Adventuring Gear',
+      url: '/api/equipment-categories/adventuring-gear',
     },
     {
-      index: 'dex',
-      name: 'DEX',
-      url: '/api/ability-scores/dex',
+      index: 'armor',
+      name: 'Armor',
+      url: '/api/equipment-categories/armor',
     },
     {
-      index: 'con',
-      name: 'CON',
-      url: '/api/ability-scores/con',
+      index: 'artisans-tools',
+      name: "Artisan's Tools",
+      url: '/api/equipment-categories/artisans-tools',
     },
   ];
   const request = mockRequest({ query: {} });
 
   it('returns a list of objects', async () => {
-    mockingoose(AbilityScore).toReturn(findDoc, 'find');
+    mockingoose(EquipmentCategory).toReturn(findDoc, 'find');
 
-    await simpleController.index(request, response, mockNext);
+    await EquipmentCategoryController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -42,9 +40,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(AbilityScore).toReturn(error, 'find');
+      mockingoose(EquipmentCategory).toReturn(error, 'find');
 
-      await simpleController.index(request, response, mockNext);
+      await EquipmentCategoryController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -55,18 +53,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'str',
-    name: 'STR',
-    url: '/api/ability-scores/str',
+    index: 'adventuring-gear',
+    name: 'Adventuring Gear',
+    url: '/api/equipment-categories/adventuring-gear',
   };
 
-  const showParams = { index: 'str' };
+  const showParams = { index: 'adventuring-gear' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(AbilityScore).toReturn(findOneDoc, 'findOne');
+    mockingoose(EquipmentCategory).toReturn(findOneDoc, 'findOne');
 
-    await simpleController.show(request, response, mockNext);
+    await EquipmentCategoryController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -74,11 +72,11 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(AbilityScore).toReturn(null, 'findOne');
+      mockingoose(EquipmentCategory).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await simpleController.show(invalidRequest, response, mockNext);
+      await EquipmentCategoryController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -89,9 +87,9 @@ describe('show', () => {
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(AbilityScore).toReturn(error, 'findOne');
+      mockingoose(EquipmentCategory).toReturn(error, 'findOne');
 
-      await simpleController.show(request, response, mockNext);
+      await EquipmentCategoryController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
