+++ src/tests/controllers/api/magicItemController.test.ts
@@ -2,24 +2,17 @@ const mockingoose = require('mockingoose');
 jest.mock('redis', () => {
   const redis = require('redis-mock');
   return redis;
-});
-const redis = require('redis');
-const { mockRequest, mockResponse, mockNext } = require('../../support/requestHelpers');
-const { redisClient } = require('../../../util');
-const MagicItem = require('../../../models/magicItem');
-const MagicItemController = require('../../../controllers/api/magicItemController');
+})
+import redis from 'redis';
+import { mockRequest, mockResponse, mockNext } from '../../support/requestHelpers';
+import { redisClient } from '../../../util';
+import MagicItem from '../../../models/magicItem';
+import * as MagicItemController from '../../../controllers/api/magicItemController';
 
-let response;
+let response: any;
 beforeEach(() => {
-  const client = redis.createClient();
-  // TODO: redis-mock does not support redis@4.0.0
-  client.flushall();
   mockingoose.resetAll();
   response = mockResponse();
-});
-
-afterAll(() => {
-  redisClient.quit();
 });
 
 describe('index', () => {
