+++ src/tests/controllers/api/spellController.test.ts
@@ -3,23 +3,16 @@ jest.mock('redis', () => {
   const redis = require('redis-mock');
   return redis;
 });
-const redis = require('redis');
-const { mockRequest, mockResponse, mockNext } = require('../../support/requestHelpers');
-const { redisClient } = require('../../../util');
-const Spell = require('../../../models/spell');
-const SpellController = require('../../../controllers/api/spellController');
+import redis from 'redis';
+import { mockRequest, mockResponse, mockNext } from '../../support/requestHelpers';
+import { redisClient } from '../../../util';
+import Spell from '../../../models/spell';
+import * as SpellController from '../../../controllers/api/spellController';
 
-let response;
+let response: any;
 beforeEach(() => {
-  const client = redis.createClient();
-  // TODO: redis-mock does not support redis@4.0.0
-  client.flushall();
   mockingoose.resetAll();
   response = mockResponse();
-});
-
-afterAll(() => {
-  redisClient.quit();
 });
 
 describe('index', () => {
