+++ src/tests/controllers/api/traitController.test.ts
@@ -1,9 +1,9 @@
 const mockingoose = require('mockingoose');
-const { mockRequest, mockResponse, mockNext } = require('../../support/requestHelpers');
-const Proficiency = require('../../../models/proficiency');
-const ProficiencyController = require('../../../controllers/api/proficiencyController');
+import { mockRequest, mockResponse, mockNext } from '../../support/requestHelpers';
+import Trait from '../../../models/trait';
+import TraitController from '../../../controllers/api/traitController';
 
-let response;
+let response: any;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
@@ -12,27 +12,27 @@ beforeEach(() => {
 describe('index', () => {
   const findDoc = [
     {
-      index: 'alchemists-supplies',
-      name: "Alchemist's supplies",
-      url: '/api/proficiencies/alchemists-supplies',
+      index: 'artificers-lore',
+      name: "Artificer's Lore",
+      url: '/api/traits/artificers-lore',
     },
     {
-      index: 'all-armor',
-      name: 'All armor',
-      url: '/api/proficiencies/all-armor',
+      index: 'brave',
+      name: 'Brave',
+      url: '/api/traits/brave',
     },
     {
-      index: 'bagpipes',
-      name: 'Bagpipes',
-      url: '/api/proficiencies/bagpipes',
+      index: 'breath-weapon',
+      name: 'Breath Weapon',
+      url: '/api/traits/breath-weapon',
     },
   ];
   const request = mockRequest({ query: {} });
 
   it('returns a list of objects', async () => {
-    mockingoose(Proficiency).toReturn(findDoc, 'find');
+    mockingoose(Trait).toReturn(findDoc, 'find');
 
-    await ProficiencyController.index(request, response, mockNext);
+    await TraitController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -40,9 +40,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Proficiency).toReturn(error, 'find');
+      mockingoose(Trait).toReturn(error, 'find');
 
-      await ProficiencyController.index(request, response, mockNext);
+      await TraitController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -53,18 +53,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'alchemists-supplies',
-    name: "Alchemist's supplies",
-    url: '/api/proficiencies/alchemists-supplies',
+    index: 'abyssal',
+    name: 'Abyssal',
+    url: '/api/traits/abyssal',
   };
 
-  const showParams = { index: 'alchemists-supplies' };
+  const showParams = { index: 'abyssal' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(Proficiency).toReturn(findOneDoc, 'findOne');
+    mockingoose(Trait).toReturn(findOneDoc, 'findOne');
 
-    await ProficiencyController.show(request, response, mockNext);
+    await TraitController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -72,11 +72,11 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(Proficiency).toReturn(null, 'findOne');
+      mockingoose(Trait).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await ProficiencyController.show(invalidRequest, response, mockNext);
+      await TraitController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -87,9 +87,9 @@ describe('show', () => {
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Proficiency).toReturn(error, 'findOne');
+      mockingoose(Trait).toReturn(error, 'findOne');
 
-      await ProficiencyController.show(request, response, mockNext);
+      await TraitController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
