+++ src/tests/controllers/api/weaponPropertyController.test.ts
@@ -1,9 +1,9 @@
 const mockingoose = require('mockingoose');
-const { mockRequest, mockResponse, mockNext } = require('../../support/requestHelpers');
-const Equipment = require('../../../models/equipment');
-const EquipmentController = require('../../../controllers/api/equipmentController');
+import { mockRequest, mockResponse, mockNext } from '../../support/requestHelpers';
+import WeaponProperty from '../../../models/weaponProperty';
+import WeaponPropertyController from '../../../controllers/api/weaponPropertyController';
 
-let response;
+let response: any;
 beforeEach(() => {
   mockingoose.resetAll();
   response = mockResponse();
@@ -12,27 +12,27 @@ beforeEach(() => {
 describe('index', () => {
   const findDoc = [
     {
-      index: 'abacus',
-      name: 'Abacus',
-      url: '/api/equipment/abacus',
+      index: 'ammunition',
+      name: 'Ammunition',
+      url: '/api/weapon-properties/ammunition',
     },
     {
-      index: 'acid-vial',
-      name: 'Acid (vial)',
-      url: '/api/equipment/acid-vial',
+      index: 'finesse',
+      name: 'Finesse',
+      url: '/api/weapon-properties/finesse',
     },
     {
-      index: 'alchemists-fire-flask',
-      name: "Alchemist's fire (flask)",
-      url: '/api/equipment/alchemists-fire-flask',
+      index: 'heavy',
+      name: 'Heavy',
+      url: '/api/weapon-properties/heavy',
     },
   ];
   const request = mockRequest({ query: {} });
 
   it('returns a list of objects', async () => {
-    mockingoose(Equipment).toReturn(findDoc, 'find');
+    mockingoose(WeaponProperty).toReturn(findDoc, 'find');
 
-    await EquipmentController.index(request, response, mockNext);
+    await WeaponPropertyController.index(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
   });
@@ -40,9 +40,9 @@ describe('index', () => {
   describe('when something goes wrong', () => {
     it('handles the error', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Equipment).toReturn(error, 'find');
+      mockingoose(WeaponProperty).toReturn(error, 'find');
 
-      await EquipmentController.index(request, response, mockNext);
+      await WeaponPropertyController.index(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -53,18 +53,18 @@ describe('index', () => {
 
 describe('show', () => {
   const findOneDoc = {
-    index: 'abacus',
-    name: 'Abacus',
-    url: '/api/equipment/abacus',
+    index: 'ammunition',
+    name: 'Ammunition',
+    url: '/api/weapon-properties/ammunition',
   };
 
-  const showParams = { index: 'abacus' };
+  const showParams = { index: 'ammunition' };
   const request = mockRequest({ params: showParams });
 
   it('returns an object', async () => {
-    mockingoose(Equipment).toReturn(findOneDoc, 'findOne');
+    mockingoose(WeaponProperty).toReturn(findOneDoc, 'findOne');
 
-    await EquipmentController.show(request, response, mockNext);
+    await WeaponPropertyController.show(request, response, mockNext);
 
     expect(response.status).toHaveBeenCalledWith(200);
     expect(response.json).toHaveBeenCalledWith(expect.objectContaining(showParams));
@@ -72,11 +72,11 @@ describe('show', () => {
 
   describe('when the record does not exist', () => {
     it('404s', async () => {
-      mockingoose(Equipment).toReturn(null, 'findOne');
+      mockingoose(WeaponProperty).toReturn(null, 'findOne');
 
       const invalidShowParams = { index: 'abcd' };
       const invalidRequest = mockRequest({ params: invalidShowParams });
-      await EquipmentController.show(invalidRequest, response, mockNext);
+      await WeaponPropertyController.show(invalidRequest, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
@@ -87,9 +87,9 @@ describe('show', () => {
   describe('when something goes wrong', () => {
     it('is handled', async () => {
       const error = new Error('Something went wrong');
-      mockingoose(Equipment).toReturn(error, 'findOne');
+      mockingoose(WeaponProperty).toReturn(error, 'findOne');
 
-      await EquipmentController.show(request, response, mockNext);
+      await WeaponPropertyController.show(request, response, mockNext);
 
       expect(response.status).not.toHaveBeenCalled();
       expect(response.json).not.toHaveBeenCalled();
